+++ scripts/google/snapd/run_google_sru.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
 . "$SCRIPTS_DIR/utils/snap_info.sh"
-sudo apt install -y jq unzip
+if ! dpkg -l jq unzip >/dev/null; then
+	sudo apt install -y jq unzip
+fi
+
 if [ "$BRANCH" = beta ]; then
 	BRANCH=$(get_beta_branch "$ARCH")
 elif [ "$BRANCH" = edge ]; then
